requirements.txt
altair
numpy
pandas
pandas-datareader
Pillow
plotly
scipy
scikit-learn>=0.18
streamlit
yfinance==0.2.40